src/drt/src/global.h is very generic; rename
`src/drt/src/global.h` is often used witout any prefix
at all (so just included as "global.h"), which is conflicing
if there happens to be e.g. `@abc` in the include path (which also
has a `"global.h"`. It also makes the current cyclic dependencies
to fix a bit more confusing.

Rename to `drt-global.h` to have an unambiguous name.
