+++ src/drt/src/GraphicsFactory.cpp
@@ -8,9 +8,9 @@
 
 #include "dr/AbstractDRGraphics.h"
 #include "dr/FlexDR_graphics.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "gui/gui.h"
 #include "odb/db.h"
 #include "pa/AbstractPAGraphics.h"
