+++ src/drt/src/TritonRoute.cpp
@@ -35,6 +35,7 @@
 #include "distributed/frArchive.h"
 #include "dr/AbstractDRGraphics.h"
 #include "dr/FlexDR.h"
+#include "drt-global.h"
 #include "drt/PinAccessService.h"
 #include "dst/Distributed.h"
 #include "dst/JobMessage.h"
@@ -43,7 +44,6 @@
 #include "frProfileTask.h"
 #include "frRTree.h"
 #include "gc/FlexGC.h"
-#include "global.h"
 #include "io/GuideProcessor.h"
 #include "io/io.h"
 #include "odb/db.h"
