+++ src/drt/src/distributed/RoutingCallBack.h
@@ -20,11 +20,11 @@
 #include "distributed/frArchive.h"
 #include "distributed/paUpdate.h"
 #include "dr/FlexDR.h"
+#include "drt-global.h"
 #include "drt/TritonRoute.h"
 #include "dst/Distributed.h"
 #include "dst/JobCallBack.h"
 #include "dst/JobMessage.h"
-#include "global.h"
 #include "omp.h"
 #include "pa/FlexPA.h"
 #include "utl/Logger.h"
