+++ src/drt/src/dr/FlexWavefront.h
@@ -9,8 +9,8 @@
 
 #include "db/infra/frBox.h"
 #include "dr/FlexMazeTypes.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
-#include "global.h"
 
 namespace drt {
 class FlexWavefrontGrid
