+++ src/drt/src/drt-global.h
