+++ src/drt/src/frRegionQuery.cpp
@@ -14,10 +14,10 @@
 #include "db/obj/frBlockObject.h"
 #include "db/obj/frBlockage.h"
 #include "db/obj/frInstBlockage.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
 #include "frRTree.h"
-#include "global.h"
 #include "odb/dbTransform.h"
 #include "utl/Logger.h"
 #include "utl/algorithms.h"
