+++ src/drt/src/gc/FlexGC.h
@@ -11,9 +11,9 @@
 #include "db/obj/frBlockObject.h"
 #include "db/obj/frMarker.h"
 #include "db/tech/frTechObject.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "utl/Logger.h"
 
 namespace drt {
