+++ src/drt/src/gc/FlexGC_cut.cpp
@@ -13,11 +13,11 @@
 #include "db/obj/frMarker.h"
 #include "db/obj/frVia.h"
 #include "db/tech/frConstraint.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frProfileTask.h"
 #include "gc/FlexGC.h"
 #include "gc/FlexGC_impl.h"
-#include "global.h"
 #include "odb/db.h"
 #include "odb/dbTypes.h"
 #include "odb/geom.h"
