+++ src/drt/src/gc/FlexGC_impl.h
@@ -20,10 +20,10 @@
 #include "db/tech/frLayer.h"
 #include "db/tech/frTechObject.h"
 #include "dr/FlexDR.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
 #include "gc/FlexGC.h"
-#include "global.h"
 #include "odb/db.h"
 
 namespace odb {
