+++ src/drt/src/io/GuideProcessor.cpp
@@ -24,10 +24,10 @@
 #include "db/obj/frBlockObject.h"
 #include "db/obj/frGCellPattern.h"
 #include "db/tech/frLayer.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
 #include "frProfileTask.h"
-#include "global.h"
 #include "odb/db.h"
 #include "odb/dbTypes.h"
 #include "odb/geom.h"
