+++ src/drt/src/io/io_parser_helper.cpp
@@ -16,9 +16,9 @@
 #include "db/obj/frFig.h"
 #include "db/obj/frVia.h"
 #include "db/tech/frConstraint.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frProfileTask.h"
-#include "global.h"
 #include "io/io.h"
 #include "odb/dbTransform.h"
 #include "odb/dbTypes.h"
