+++ src/drt/src/pa/FlexPA.cpp
@@ -22,12 +22,12 @@
 #include "distributed/PinAccessJobDescription.h"
 #include "distributed/frArchive.h"
 #include "distributed/paUpdate.h"
+#include "drt-global.h"
 #include "dst/Distributed.h"
 #include "dst/JobMessage.h"
 #include "frBaseTypes.h"
 #include "frProfileTask.h"
 #include "gc/FlexGC.h"
-#include "global.h"
 #include "odb/db.h"
 #include "omp.h"
 #include "pa/AbstractPAGraphics.h"
