+++ src/drt/src/pa/FlexPA.h
@@ -25,9 +25,9 @@
 #include "db/tech/frLayer.h"
 #include "db/tech/frTechObject.h"
 #include "db/tech/frViaDef.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "odb/db.h"
 #include "pa/FlexPA_unique.h"
 namespace gtl = boost::polygon;
