+++ src/drt/src/pa/FlexPA_graphics.cpp
@@ -11,9 +11,9 @@
 
 #include "db/obj/frBlockObject.h"
 #include "db/obj/frMPin.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "gui/gui.h"
 #include "pa/FlexPA.h"
 #include "pa/FlexPA_unique.h"
