+++ src/drt/src/pa/FlexPA_graphics.h
@@ -10,8 +10,8 @@
 
 #include "db/obj/frBlockObject.h"
 #include "db/obj/frMPin.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
-#include "global.h"
 #include "gui/gui.h"
 #include "pa/AbstractPAGraphics.h"
 #include "pa/FlexPA.h"
