+++ src/drt/src/pa/FlexPA_unique.h
@@ -16,9 +16,9 @@
 #include "db/obj/frMPin.h"
 #include "db/obj/frTrackPattern.h"
 #include "db/tech/frTechObject.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "odb/db.h"
 #include "odb/dbTypes.h"
