+++ src/drt/src/rp/FlexRP.h
@@ -10,9 +10,9 @@
 #include "db/tech/frLayer.h"
 #include "db/tech/frTechObject.h"
 #include "db/tech/frViaDef.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "utl/Logger.h"
 
 namespace drt {
