+++ src/drt/src/serialization.h
@@ -35,9 +35,9 @@
 #include "db/obj/frVia.h"
 #include "distributed/drUpdate.h"
 #include "distributed/paUpdate.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
-#include "global.h"
 #include "odb/dbTransform.h"
 #include "odb/dbTypes.h"
 #include "odb/geom.h"
