+++ src/drt/src/ta/FlexTA.h
@@ -16,10 +16,10 @@
 #include "db/tech/frConstraint.h"
 #include "db/tech/frTechObject.h"
 #include "db/tech/frViaDef.h"
+#include "drt-global.h"
 #include "frBaseTypes.h"
 #include "frDesign.h"
 #include "frRegionQuery.h"
-#include "global.h"
 #include "utl/Logger.h"
 
 namespace drt {
