+++ src/drt/test/fixture.cpp
@@ -27,7 +27,7 @@
 #include "frRegionQuery.h"
 #include "odb/db.h"
 #include "odb/dbTypes.h"
-#include "src/drt/src/global.h"
+#include "src/drt/src/drt-global.h"
 #include "utl/Logger.h"
 
 using odb::dbTechLayerDir;
